requirements.txt
streamlit
deepface
tensorflow==2.12.0
opencv-python-headless
numpy
pillow











